Move commands to testing helper
