+++ e2e/kube/e2ehelper/e2e_helper.go
@@ -1,11 +1,8 @@
 package e2ehelper
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"os"
-	"os/exec"
 	"strconv"
 	"strings"
 
@@ -13,13 +10,11 @@ import (
 	"github.com/pkg/errors"
 
 	"code.cloudfoundry.org/cf-operator/integration/environment"
+	"code.cloudfoundry.org/cf-operator/testing"
 )
 
 const (
-	CFOperatorRelease    = "cf-operator"
 	installTimeOutInSecs = "600"
-	helmCmd              = "helm"
-	kubeCtlCmd           = "kubectl"
 )
 
 var (
@@ -35,12 +30,12 @@ func SetUpEnvironment(chartPath string) (string, environment.TearDownFunc, error
 
 	// Ensure tiller is there, if not
 	// then create it, via "init"
-	err := RunHelmBinaryWithCustomErr(helmCmd, "version", "-s")
+	err := testing.RunHelmBinaryWithCustomErr("version", "-s")
 	if err != nil {
 		switch err := err.(type) {
-		case *CustomError:
+		case *testing.CustomError:
 			if strings.Contains(err.StdOut, "could not find tiller") {
-				_, err := environment.RunBinary(helmCmd, "init", "--wait")
+				err := testing.RunHelmBinaryWithCustomErr("init", "--wait")
 				if err != nil {
 					return "", nil, err
 				}
@@ -62,8 +57,8 @@ func SetUpEnvironment(chartPath string) (string, environment.TearDownFunc, error
 	fmt.Println("Setting up in test namespace '" + namespace + "'...")
 
 	// TODO: find relative path here
-	_, err = environment.RunBinary(helmCmd, "install", chartPath,
-		"--name", fmt.Sprintf("%s-%s", CFOperatorRelease, namespace),
+	err = testing.RunHelmBinaryWithCustomErr("install", chartPath,
+		"--name", fmt.Sprintf("%s-%s", testing.CFOperatorRelease, namespace),
 		"--namespace", namespace,
 		"--timeout", installTimeOutInSecs,
 
@@ -75,22 +70,22 @@ func SetUpEnvironment(chartPath string) (string, environment.TearDownFunc, error
 
 	teardownFunc := func() error {
 		var messages string
-		err = DeleteWebhookSecret(namespace)
+		err = testing.DeleteSecret(namespace, "cf-operator-webhook-server-cert")
 		if err != nil {
 			messages = fmt.Sprintf("%v%v\n", messages, err.Error())
 		}
 
-		err = environment.DeleteWebhooks(namespace, kubeCtlCmd)
+		err = testing.DeleteWebhooks(namespace)
 		if err != nil {
 			messages = fmt.Sprintf("%v%v\n", messages, err.Error())
 		}
 
-		_, err := environment.RunBinary(helmCmd, "delete", fmt.Sprintf("%s-%s", CFOperatorRelease, namespace), "--purge")
+		err := testing.RunHelmBinaryWithCustomErr("delete", fmt.Sprintf("%s-%s", testing.CFOperatorRelease, namespace), "--purge")
 		if err != nil {
 			messages = fmt.Sprintf("%v%v\n", messages, err.Error())
 		}
 
-		err = environment.DeleteNamespace(namespace, kubeCtlCmd)
+		err = testing.DeleteNamespace(namespace)
 		if err != nil {
 			messages = fmt.Sprintf("%v%v\n", messages, err.Error())
 		}
@@ -108,16 +103,11 @@ func SetUpEnvironment(chartPath string) (string, environment.TearDownFunc, error
 
 // ClusterCrdsExist verify if cf-operator CRDs are already in place
 func ClusterCrdsExist() (bool, error) {
-	customResource := &ClusterCrd{}
-	stdOutput, err := environment.RunBinary(kubeCtlCmd, "get", "crds", "-o=json")
+	customResource, err := testing.GetCRDs()
 	if err != nil {
 		return false, err
 	}
 
-	d := json.NewDecoder(bytes.NewReader(stdOutput))
-	if err := d.Decode(customResource); err != nil {
-		return false, err
-	}
 	crds := []string{"boshdeployments.fissile.cloudfoundry.org",
 		"extendedjobs.fissile.cloudfoundry.org",
 		"extendedsecrets.fissile.cloudfoundry.org",
@@ -125,24 +115,14 @@ func ClusterCrdsExist() (bool, error) {
 
 	if len(customResource.Items) > 0 {
 		for _, crdName := range crds {
-			if !ContainsElement(customResource, crdName) {
-				return false, fmt.Errorf("missing CRDs in cluster, %s not found", crdName)
+			if !customResource.ContainsElement(crdName) {
+				return false, fmt.Errorf("missing CRD in cluster CRDs %+v, %s not found", customResource.Items, crdName)
 			}
 		}
 
 		return true, nil
 	}
 	return false, errors.New("Missing CRDs in cluster")
-}
-
-// RunHelmBinaryWithCustomErr executes a desire binary
-func RunHelmBinaryWithCustomErr(binaryName string, args ...string) error {
-	cmd := exec.Command(binaryName, args...)
-	stdOutput, err := cmd.CombinedOutput()
-	if err != nil {
-		return &CustomError{strings.Join(cmd.Args, " "), string(stdOutput), err}
-	}
-	return nil
 }
 
 // CreateTestNamespace generates a namespace based on a env variable
@@ -154,50 +134,10 @@ func CreateTestNamespace() (string, error) {
 	namespace := prefix + "-" + strconv.Itoa(config.GinkgoConfig.ParallelNode) + "-" + strconv.Itoa(int(nsIndex))
 	nsIndex++
 
-	_, err := environment.RunBinary(kubeCtlCmd, "create", "ns", namespace)
+	err := testing.CreateNamespace(namespace)
 	if err != nil {
 		return "", err
 	}
 
 	return namespace, nil
-}
-
-// DeleteWebhookSecret removes left overs from the cf-operator chart,
-// like the webhook server certificate secret in a specific namespace
-func DeleteWebhookSecret(ns string) error {
-	_, err := environment.RunBinary(kubeCtlCmd, "--namespace", ns, "delete", "secret", "cf-operator-webhook-server-cert", "--ignore-not-found")
-
-	return err
-}
-
-// CustomError containing stdOuput of a binary execution
-type CustomError struct {
-	Msg    string
-	StdOut string
-	Err    error
-}
-
-func (e *CustomError) Error() string {
-	return fmt.Sprintf("%s:%v:%v", e.Msg, e.Err, e.StdOut)
-}
-
-// ClusterCrd defines a list of CRDs
-type ClusterCrd struct {
-	Items []struct {
-		APIVersion string `json:"apiVersion"`
-		Kind       string `json:"kind"`
-		Metadata   struct {
-			Name string `json:"name"`
-		} `json:"metadata"`
-	} `json:"items"`
-}
-
-// ContainsElement verify if a CRD exist
-func ContainsElement(list *ClusterCrd, element string) bool {
-	for _, n := range list.Items {
-		if n.Metadata.Name == element {
-			return true
-		}
-	}
-	return false
 }
